Finished SocketComms; untested
